динамическое имя папки с файлами

+++ main.py
@@ -7,7 +7,7 @@ import shutil
 
 if __name__ == "__main__":
 
-    if IS_DELETE_TMP:
+    if False and IS_DELETE_TMP:
         if os.path.isdir(PATH_LOG):
             shutil.rmtree(PATH_LOG)
         if os.path.isdir(PATH_OUTPUT):
@@ -18,6 +18,6 @@ if __name__ == "__main__":
     args = getArgs()
     namespace = args.parse_args(sys.argv[1:])
     parser = Parser(file_name=namespace.name, inn=namespace.inn, file_config=namespace.config,
-                    union=namespace.union, path_down=PATH_OUTPUT, hash=namespace.hash,
+                    union=namespace.union, path_down=PATH_DOWNLOAD, hash=namespace.hash,
                     file_down=f'output_{datetime.now().strftime("%Y-%m-%d_%H-%M-%S")}.zip')
     parser.start()
